Added more resources
It's in beta so it's not finished yet but if you would like to check it out I added much more to resources.

I also added them in divisions so the contrast of the web font would not be tangled with the gradient of the website.

A note would be to have a table for transferable courses